refactor: FilesView to MyFilesModal
